Propagate exceptions to the main thread
With this commit, exceptions occuring from a worker thread are
propagated to the main thread, which re-throw them.

As a consequence, any unhandled exception during processing will
interrupt the whole process. I think this is the desired behaviour as
expected exceptions should be handled from within the worker threads.

The main benefit is that users now see the error and can remediate them,
notably ones thrown because of disassembler misconfiguration e.g.
BinExport plugin not installed...

Signed-off-by: lbuckwell <lbuckwell@quarkslab.com>